+++ CHANGELOG.md
@@ -37,6 +37,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Updated CDI supported version from `5.2` to `5.3`
 -   Changed `AccountInfo` to `AccountInfoInput` in various methods for better type safety. This is required in order to allow querying by a single WebAuthn `credentialId`, while the WebAuthn login method contains an array of `credentialId`
 
+## [21.1.1] - 2025-03-06
+
+-   Fixes an issue where the response body was not being cloned when using cache
+-   Fixes type of `cookies` to `string[]` instead of `string` in:
+    -   Return type of `authorization` in `RecipeInterface` in `oauth2provider` recipe
+    -   Return type of `authGET` in `APIInterface` in `oauth2provider` recipe
+    -   Return type of `loginGET` in `APIInterface` in `oauth2provider` recipe
+
 ## [21.1.0] - 2024-11-19
 
 -   Adds `getCookieNameForTokenType` config option to allow customizing the cookie name for a token type.